requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0.1
openai>=1.40.0
pydantic>=2.6.0
